+++ src/main/org/firebirdsql/gds/ng/FbConnectionProperties.java
@@ -174,7 +174,8 @@ public final class FbConnectionProperties extends AbstractAttachProperties<IConn
     @Deprecated
     public void fromDpb(DatabaseParameterBuffer dpb) throws SQLException {
         for (Parameter parameter : dpb) {
-            switch (parameter.getType()) {
+            final int parameterType = parameter.getType();
+            switch (parameterType) {
             case isc_dpb_user_name:
                 setUser(parameter.getValueAsString());
                 break;
@@ -234,7 +235,11 @@ public final class FbConnectionProperties extends AbstractAttachProperties<IConn
             case isc_dpb_specific_auth_data:
                 break;
             default:
-                log.warn(String.format("Unknown or unsupported parameter with type %d added to extra database parameters", parameter.getType()));
+                if (parameterType < jaybirdMinIscDpbValue || parameterType > jaybirdMaxIscDpbValue) {
+                    log.warn(String.format(
+                            "Unknown or unsupported parameter with type %d added to extra database parameters",
+                            parameterType));
+                }
                 parameter.copyTo(getExtraDatabaseParameters(), null);
                 dirtied();
             }