+++ src/main/org/firebirdsql/jdbc/FBDatabaseMetaData.java
@@ -3094,8 +3094,9 @@ public class FBDatabaseMetaData implements FirebirdDatabaseMetaData {
 
     @Override
     public boolean supportsGetGeneratedKeys() throws SQLException {
-        return AbstractGeneratedKeysQuery.isGeneratedKeysSupportLoaded()
-                && firebirdSupportInfo.supportsInsertReturning();
+        return firebirdSupportInfo.supportsInsertReturning()
+                && GeneratedKeysSupportFactory.isGeneratedKeysSupportLoaded()
+                && connection.getGeneratedKeysSupport().supportsGetGeneratedKeys();
     }
 
     /**